backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
deepface
pillow
numpy
opencv-python-headless
tf-keras